image-forgery-detection/requirements.txt
keras
matplotlib
numpy
pillow
pyqt5
seaborn
sklearn
tensorflow
tqdm
flask